Handle Reused SAP Numbers in Vacancy Controller

+++ database/migrations/2025_02_19_111238_add_auto_deleted_column_to_vacancies_table.php
@@ -4,16 +4,17 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up()
     {
-        Schema::table('vacancies', function (Blueprint $table) {
-            $table->enum('auto_deleted', ['Yes', 'No'])->default('No')->nullable()->after('deleted');
-        });
+        if (!Schema::hasColumn('vacancies', 'auto_deleted')) {
+            Schema::table('vacancies', function (Blueprint $table) {
+                $table->enum('auto_deleted', ['Yes', 'No'])->default('No')->nullable()->after('deleted');
+            });
+        }
     }
 
     /**
@@ -21,8 +22,10 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('vacancies', function (Blueprint $table) {
-            $table->dropColumn('auto_deleted');
-        });
+        if (Schema::hasColumn('vacancies', 'auto_deleted')) {
+            Schema::table('vacancies', function (Blueprint $table) {
+                $table->dropColumn('auto_deleted');
+            });
+        }
     }
 };
